requirements.txt
discord.py~=1.7.3
confuse~=1.4.0
